+++ client/web/src/enterprise/codeintel/badge/components/CodeIntelBadge.module.scss
@@ -2,10 +2,6 @@
     min-width: 20rem;
 }
 
-.language-request {
-    font-weight: normal;
-}
-
 .badge-multiple {
     min-width: 7rem;
 }
