+++ client/web/src/enterprise/codeintel/badge/components/InternalBadgeContent.tsx
@@ -1,0 +1,121 @@
+import React from 'react'
+
+import { isErrorLike } from '@sourcegraph/common'
+import { SettingsCascadeProps } from '@sourcegraph/shared/src/settings/settings'
+import { MenuDivider } from '@sourcegraph/wildcard'
+
+import { Collapsible } from '../../../../components/Collapsible'
+import { Timestamp } from '../../../../components/time/Timestamp'
+import { CodeIntelIndexer } from '../../shared/components/CodeIntelIndexer'
+import { UseCodeIntelStatusPayload } from '../hooks/useCodeIntelStatus'
+
+import { UploadOrIndexMetaTable } from './UploadOrIndexMetaTable'
+
+export type InternalBadgeContentProps = SettingsCascadeProps & {
+    data: UseCodeIntelStatusPayload
+    now?: () => Date
+}
+
+export const InternalBadgeContent: React.FunctionComponent<InternalBadgeContentProps> = ({
+    data,
+    now,
+    settingsCascade,
+}) => {
+    const forNerds =
+        !isErrorLike(settingsCascade.final) &&
+        settingsCascade.final?.experimentalFeatures?.codeIntelRepositoryBadge?.forNerds
+
+    if (!forNerds) {
+        return null
+    }
+
+    const preciseSupportLevels = [...new Set((data.preciseSupport || []).map(support => support.supportLevel))].sort()
+    const searchBasedSupportLevels = [
+        ...new Set((data?.searchBasedSupport || []).map(support => support.supportLevel)),
+    ].sort()
+
+    return (
+        <>
+            <MenuDivider />
+
+            <div className="px-2 py-1">
+                <Collapsible titleAtStart={true} title={<h3>Activity (repo)</h3>}>
+                    <div>
+                        <span>
+                            Last auto-indexing job schedule attempt:{' '}
+                            {data.lastIndexScan ? <Timestamp date={data.lastIndexScan} now={now} /> : <>never</>}
+                        </span>
+                    </div>
+                    <div>
+                        <span>
+                            Last upload retention scan:{' '}
+                            {data.lastUploadRetentionScan ? (
+                                <Timestamp date={data.lastUploadRetentionScan} now={now} />
+                            ) : (
+                                <>never</>
+                            )}
+                        </span>
+                    </div>
+                </Collapsible>
+
+                <Collapsible titleAtStart={true} title={<h3>Support (tree)</h3>}>
+                    <ul>
+                        {preciseSupportLevels.map(supportLevel => (
+                            <li key={`precise-support-level-${supportLevel}`}>
+                                <code>{supportLevel}</code>
+                                <ul>
+                                    {data.preciseSupport
+                                        ?.filter(support => support.supportLevel === supportLevel)
+                                        .map(support =>
+                                            support.indexers?.map(indexer => (
+                                                <li key={`precise-support-level-${supportLevel}-${indexer.name}`}>
+                                                    <CodeIntelIndexer indexer={indexer} /> (
+                                                    {support.confidence && (
+                                                        <span className="text-muted">{support.confidence}</span>
+                                                    )}
+                                                    )
+                                                </li>
+                                            ))
+                                        )}
+                                </ul>
+                            </li>
+                        ))}
+
+                        {searchBasedSupportLevels.map(supportLevel => (
+                            <li key={`search-support-level-${supportLevel}`}>
+                                <code>{supportLevel}</code>
+                                <ul>
+                                    {data.searchBasedSupport
+                                        ?.filter(support => support.supportLevel === supportLevel)
+                                        .map(support => (
+                                            <li key={`search-support-level-${supportLevel}-${support.language}`}>
+                                                {support.language}
+                                            </li>
+                                        ))}
+                                </ul>
+                            </li>
+                        ))}
+                    </ul>
+                </Collapsible>
+
+                <Collapsible titleAtStart={true} title={<h3>Recent uploads (repo)</h3>}>
+                    <UploadOrIndexMetaTable
+                        prefix="recent-uploads"
+                        nodes={data.recentUploads.flatMap(namespacedUploads => namespacedUploads.uploads)}
+                    />
+                </Collapsible>
+
+                <Collapsible titleAtStart={true} title={<h3>Recent indexes (repo)</h3>}>
+                    <UploadOrIndexMetaTable
+                        prefix="recent-indexes"
+                        nodes={data.recentIndexes.flatMap(namespacedIndexes => namespacedIndexes.indexes)}
+                    />
+                </Collapsible>
+
+                <Collapsible titleAtStart={true} title={<h3>Uploads providing intel (tree)</h3>}>
+                    <UploadOrIndexMetaTable prefix="active-uploads" nodes={data.activeUploads} />
+                </Collapsible>
+            </div>
+        </>
+    )
+}
