+++ client/web/src/enterprise/codeintel/badge/components/RequestLink.module.scss
@@ -1,0 +1,3 @@
+.language-request {
+    font-weight: normal;
+}
